@@ -85,7 +85,7 @@ namespace MaterialDesignDemo.Domain
 
         public ICommand Open { get; }
 
-        private void Execute(object o)
+        private void Execute(object? _)
         {
             Link.OpenInBrowser(Url);
         }
