@@ -5,8 +5,9 @@ namespace MaterialDesignDemo.Domain
 {
     public static class Link
     {
-        public static void OpenInBrowser(string url)
+        public static void OpenInBrowser(string? url)
         {
+            if (url is null) return;
             //https://github.com/dotnet/corefx/issues/10361
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
