@@ -8,14 +8,14 @@ namespace MaterialDesign3Demo.Domain
     /// </summary>
     public class AnotherCommandImplementation : ICommand
     {
-        private readonly Action<object> _execute;
-        private readonly Func<object, bool> _canExecute;
+        private readonly Action<object?> _execute;
+        private readonly Func<object?, bool> _canExecute;
 
-        public AnotherCommandImplementation(Action<object> execute)
+        public AnotherCommandImplementation(Action<object?> execute)
             : this(execute, null)
         { }
 
-        public AnotherCommandImplementation(Action<object> execute, Func<object, bool>? canExecute)
+        public AnotherCommandImplementation(Action<object?> execute, Func<object?, bool>? canExecute)
         {
             if (execute is null) throw new ArgumentNullException(nameof(execute));
 
@@ -23,11 +23,11 @@ namespace MaterialDesign3Demo.Domain
             _canExecute = canExecute ?? (x => true);
         }
 
-        public bool CanExecute(object parameter) => _canExecute(parameter);
+        public bool CanExecute(object? parameter) => _canExecute(parameter);
 
-        public void Execute(object parameter) => _execute(parameter);
+        public void Execute(object? parameter) => _execute(parameter);
 
-        public event EventHandler CanExecuteChanged
+        public event EventHandler? CanExecuteChanged
         {
             add
             {
