@@ -3306,6 +3306,7 @@ namespace MaterialDesignThemes.Wpf
         Chip,
         IntegratedCircuit=Chip,
         Church,
+        ChurchOutline,
         Cigar,
         CigarOff,
         Circle,
@@ -10750,8 +10751,10 @@ namespace MaterialDesignThemes.Wpf
         StackOverflow,
         Stackpath,
         Stadium,
+        Arena=Stadium,
+        StadiumOutline,
+        ArenaOutline=StadiumOutline,
         StadiumVariant,
-        Arena=StadiumVariant,
         Stairs,
         StairsBox,
         StairsDown,
@@ -11074,6 +11077,13 @@ namespace MaterialDesignThemes.Wpf
         WritingSystemKatakanaHalfWidth=SyllabaryKatakanaHalfwidth,
         Symbol,
         Symfony,
+        Synagogue,
+        Shul=Synagogue,
+        Temple=Synagogue,
+        SynagogueOutline,
+        JewishOutline=SynagogueOutline,
+        ShulOutline=SynagogueOutline,
+        TempleOutline=SynagogueOutline,
         Sync,
         SyncAlert,
         SyncProblem=SyncAlert,
@@ -11273,6 +11283,10 @@ namespace MaterialDesignThemes.Wpf
         TemperatureCentigrade=TemperatureCelsius,
         TemperatureFahrenheit,
         TemperatureKelvin,
+        TempleBuddhist,
+        TempleBuddhistOutline,
+        TempleHindu,
+        TempleHinduOutline,
         Tennis,
         TennisRacket=Tennis,
         TennisRacquet=Tennis,
